+++ packages/expo-local-authentication/android/src/main/java/expo/modules/localauthentication/LocalAuthenticationModule.kt
@@ -205,11 +205,6 @@ class LocalAuthenticationModule : Module() {
       return
     }
 
-    val promptMessage = options.promptMessage
-    val promptSubtitle = options.promptSubtitle
-    val promptDescription = options.promptDescription
-    val cancelLabel = options.cancelLabel
-    val requireConfirmation = options.requireConfirmation
     val allowedAuthenticators = if (options.disableDeviceFallback) {
       options.biometricsSecurityLevel.toNativeBiometricSecurityLevel()
     } else {
@@ -219,21 +214,22 @@ class LocalAuthenticationModule : Module() {
     isAuthenticating = true
     this.promise = promise
     val executor: Executor = Executors.newSingleThreadExecutor()
-    biometricPrompt = BiometricPrompt(fragmentActivity, executor, authenticationCallback)
+    val localBiometricPrompt = BiometricPrompt(fragmentActivity, executor, authenticationCallback)
+    biometricPrompt = localBiometricPrompt
     val promptInfoBuilder = PromptInfo.Builder().apply {
-      setTitle(promptMessage)
-      setSubtitle(promptSubtitle)
-      setDescription(promptDescription)
+      setTitle(options.promptMessage)
+      setSubtitle(options.promptSubtitle)
+      setDescription(options.promptDescription)
       setAllowedAuthenticators(allowedAuthenticators)
-      if (options.disableDeviceFallback) {
-        setNegativeButtonText(cancelLabel)
+      if (options.disableDeviceFallback && options.cancelLabel != null) {
+        setNegativeButtonText(options.cancelLabel)
       }
-      setConfirmationRequired(requireConfirmation)
+      setConfirmationRequired(options.requireConfirmation)
     }
 
     val promptInfo = promptInfoBuilder.build()
     try {
-      biometricPrompt!!.authenticate(promptInfo)
+      localBiometricPrompt.authenticate(promptInfo)
     } catch (e: NullPointerException) {
       promise.reject(UnexpectedException("Canceled authentication due to an internal error", e))
     }
@@ -252,9 +248,7 @@ class LocalAuthenticationModule : Module() {
     }
 
     val promptMessage = options.promptMessage
-    val promptSubtitle = options.promptSubtitle
     val promptDescription = options.promptDescription
-    val requireConfirmation = options.requireConfirmation
 
     // BiometricPrompt callbacks are invoked on the main thread so also run this there to avoid
     // having to do locking.
@@ -273,10 +267,10 @@ class LocalAuthenticationModule : Module() {
 
       val promptInfoBuilder = PromptInfo.Builder().apply {
         setTitle(promptMessage)
-        setSubtitle(promptSubtitle)
+        setSubtitle(options.promptSubtitle)
         setDescription(promptDescription)
         setAllowedAuthenticators(BiometricManager.Authenticators.DEVICE_CREDENTIAL)
-        setConfirmationRequired(requireConfirmation)
+        setConfirmationRequired(options.requireConfirmation)
       }
 
       val promptInfo = promptInfoBuilder.build()
@@ -294,7 +288,7 @@ class LocalAuthenticationModule : Module() {
   // but it will be ignored on falling-back to device credential on biometric authentication.
   // That means, setting level to `SECURITY_LEVEL_SECRET` might be misleading for some users.
   // But there is no equivalent APIs prior to M.
-  // `andriodx.biometric.BiometricManager#canAuthenticate(int)` looks like an alternative,
+  // `androidx.biometric.BiometricManager#canAuthenticate(int)` looks like an alternative,
   // but specifying `BiometricManager.Authenticators.DEVICE_CREDENTIAL` alone is not
   // supported prior to API 30.
   // https://developer.android.com/reference/androidx/biometric/BiometricManager#canAuthenticate(int)